Display active download amount on buttons
